[feat] Added fields for ai job description in scanning history page (#91)
