+++ frontend/src/api/fetchScreenings.ts
@@ -1,30 +1,6 @@
-import type { ScreeningDetails } from "@/types/screening";
+import type { SaveScreeningRunPayload, ScreeningDetails } from "@/types/screening";
 import { API_URL } from "@/utils/variables";
 
-
-export type SaveScreeningRunPayload = {
-  title: string;
-  header: string;
-  description: string;
-  hardQualifications: string[];
-  softQualifications: string[];
-  candidates: Array<{
-    candidateId: number;
-    rank: number;
-    score: number;
-    qualified: boolean;
-    qualificationsMet: string[];
-    qualificationsMissing: string[];
-    unknowns: string[];
-    summary?: string;
-  }>;
-};
-
-export type SaveScreeningRunResponse = {
-  jobPostId: number;
-  title: string;
-  screenedAt: string;
-};
 
 export async function getScreeningHistory(): Promise<ScreeningDetails[]> {
   const response = await fetch(API_URL + "/api/results/history");
@@ -49,7 +25,7 @@ export async function getScreeningByJobPostId(
 
 export async function saveScreeningRun(
   payload: SaveScreeningRunPayload,
-): Promise<SaveScreeningRunResponse> {
+): Promise<ScreeningDetails> {
   const response = await fetch(API_URL + "/api/results/screenings", {
     method: "POST",
     headers: {
